Rename 'effects' folder to 'hooks'
