+++ src/components/Cluster/ClusterDetail/KeypairCreateModal/KeyPairCreateModal.tsx
@@ -12,8 +12,8 @@ import {
   MODAL_CHANGE_TIMEOUT,
   VALIDATION_DEBOUNCE_RATE,
 } from 'Cluster/ClusterDetail/KeypairCreateModal/Utils';
-import useDebounce from 'lib/effects/useDebounce';
 import { dedent, makeKubeConfigTextFile } from 'lib/helpers';
+import useDebounce from 'lib/hooks/useDebounce';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
 import BootstrapModal from 'react-bootstrap/lib/Modal';
