+++ src/components/GettingStarted/__tests__/CodeBlock.js
@@ -1,13 +1,13 @@
 import '@testing-library/jest-dom/extend-expect';
 
 import { fireEvent, render, waitFor } from '@testing-library/react';
-import useCopyToClipboard from 'lib/effects/useCopyToClipboard';
+import useCopyToClipboard from 'lib/hooks/useCopyToClipboard';
 import React from 'react';
 import { getComponentWithTheme } from 'testUtils/renderUtils';
 
 import { CodeBlock, Output, Prompt } from '../CodeBlock';
 
-jest.mock('lib/effects/useCopyToClipboard');
+jest.mock('lib/hooks/useCopyToClipboard');
 
 let setClipboardContentMockFn = null;
