+++ src/components/UI/RefreshableLabel.js
@@ -1,6 +1,6 @@
 import { keyframes } from '@emotion/core';
 import styled from '@emotion/styled';
-import usePrevious from 'lib/effects/usePrevious';
+import usePrevious from 'lib/hooks/usePrevious';
 import PropTypes from 'prop-types';
 import React, { useEffect, useState } from 'react';
