+++ src/components/shared/__tests__/Copyable.js
@@ -1,12 +1,12 @@
 import '@testing-library/jest-dom/extend-expect';
 
 import { fireEvent, render, waitFor } from '@testing-library/react';
-import useCopyToClipboard from 'lib/effects/useCopyToClipboard';
+import useCopyToClipboard from 'lib/hooks/useCopyToClipboard';
 import { getComponentWithTheme } from 'testUtils/renderUtils';
 
 import Copyable from '../Copyable';
 
-jest.mock('lib/effects/useCopyToClipboard');
+jest.mock('lib/hooks/useCopyToClipboard');
 
 let setClipboardContentMockFn = null;
