+++ src/lib/hooks/useDocumentTitle.ts
